Fix bug where multiple respawn packets were sent for the same event
